requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-redis==5.4.0
djangorestframework==3.16.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
idna==3.10
inflection==0.5.1
mysqlclient==2.2.7
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
